+++ neo/api/JSONRPC/JsonRpcApi.py
@@ -73,8 +73,9 @@ class JsonRpcApi:
     app = Klein()
     port = None
 
-    def __init__(self, port):
+    def __init__(self, port, wallet=None):
         self.port = port
+        self.wallet = wallet
 
     #
     # JSON-RPC API Route
@@ -244,6 +245,12 @@ class JsonRpcApi:
         elif method == "getpeers":
             return self.get_peers()
 
+        elif method == "listaddress":
+            if self.wallet:
+                return self.list_address()
+            else:
+                raise JsonRpcError(-400, "Access denied.")
+
         raise JsonRpcError.methodNotFound()
 
     def get_custom_error_payload(self, request_id, code, message):
@@ -331,4 +338,17 @@ class JsonRpcApi:
                 result['unconnected'].append({"address": addr,
                                               "port": int(port)})
 
+        return result
+
+    def list_address(self):
+        """Get information about all the addresses present on the open wallet"""
+        result = []
+        for addrStr in self.wallet.Addresses:
+            addr = self.wallet.GetAddress(addrStr)
+            result.append({
+                "address": addrStr,
+                "haskey": not addr.IsWatchOnly,
+                "label": None,
+                "watchonly": addr.IsWatchOnly,
+            })
         return result