+++ CHANGELOG.rst
@@ -10,6 +10,8 @@ All notable changes to this project are documented in this file.
   history or logs because those files are just created from scratch in the custom datadir.
 - Updated the dependencies
 - Updated MainNet bootstrap files
+- Added ``--wallet`` flag to the ``np-api-server`` command. The server can now open a wallet. `#459 <https://github.com/CityOfZion/neo-python/pull/459>`_
+- Added a partial implementation of the ``listaddress`` RPC method. `#459 <https://github.com/CityOfZion/neo-python/pull/459>`_
 
 
 [0.7.1] 2018-06-02